fix up docs and modify existing code to reflect scope change
previously I was thinking of this as a single-use job management framework backed by a remote filesystem or cloud storage. Now, I am planning to generalize it more and ditch some of the job management aspects and just make this a filesystem driver along with a remote fileserver

for now I will stick with a userspace FUSE filesystem on the client side to make hooking into gRPC and my remote server easier. If I really wanted to take it to the next level I could do the same with a low-level device controller and emulate an existing kernel-space filesystem. But this is more than enough to bite off for now