add HORREA component desktop test, fix bug
I had a bug around stream closes for PUT. I was not gracefully closing the connection, so the client was seeing an error

I really need a client library for the file PUT/GET, since it's painful to manage all that. But since there probably won't be many clients, I haven't decided if it makes more sense to keep the package local to the consumer service or put it in the HORREA directory

With the change in scope, I probably need a mechanism to do partial file reads. That will be easy to do when needed so leaving bulk file READ/WRITE for now